new file:   L2_scene.py 	modified:   batch.qsub_exec.sh 	new file:   compositeNWLR.R 	new file:   getNWLR.sh 	modified:   getRSRF.sh 	new file:   qsub_exec.NWLR.sh

+++ getRSRF.sh
@@ -19,20 +19,24 @@ getRSRF() {
     export YYYY=$(date --date="$DATE_STRING" +%Y)
     export MM=$(date --date="$DATE_STRING" +%m)
     export DD=$(date --date="$DATE_STRING" +%d)
-
-    H5FILE=$WORKDIR/GC1SG1_${YYYY}${MM}${DD}D01D_T${TILE}_L2SG_RSRFQ_2000.h5
-    FTP=ftp://ftp.gportal.jaxa.jp/standard/GCOM-C/GCOM-C.SGLI/L2.LAND.RSRF/2
+    
     #H5FILE=$WORKDIR/GC1SG1_${YYYY}${MM}${DD}D01D_T${TILE}_L2SG_RSRFQ_1001.h5
     #FTP=ftp://ftp.gportal.jaxa.jp/standard/GCOM-C/GCOM-C.SGLI/L2.LAND.RSRF/1
 
-    #H5FILE=$WORKDIR/GC1SG1_${YYYY}${MM}${DD}D01D_T${TILE}_L2SG_LTOAK_1002.h5
-    #FTP="ftp://ftp.gportal.jaxa.jp/standard/GCOM-C/GCOM-C.SGLI/L2.LAND.LTOA"
-    
-    wget -q -nc --user=heromiya --password=anonymous $FTP/$YYYY/$MM/$DD/$(basename $H5FILE) -O $H5FILE
-    for B in VN04 VN06 VN07 VN10; do # VN03 VN05 VN07 
-      	python3 h5_2_tiff.py $H5FILE Rs_${B} $H5FILE.$B.tif $GCP_INTERVAL
-	gdalwarp $WARPOPT $H5FILE.$B.tif GCOM-C/$RES/$TILE/$(basename $H5FILE).$B.tif
-    done
+    H5FILE=$WORKDIR/GC1SG1_${YYYY}${MM}${DD}D01D_T${TILE}_L2SG_RSRFQ_2000.h5
+    FTP=ftp://ftp.gportal.jaxa.jp/standard/GCOM-C/GCOM-C.SGLI/L2.LAND.RSRF/2
+
+    if [ ! -e  GCOM-C/$RES/$TILE/$(basename $H5FILE).VN04.tif ]; then
+	wget -q -nc --user=heromiya --password=anonymous $FTP/$YYYY/$MM/$DD/$(basename $H5FILE) -O $H5FILE
+	for B in VN04 VN06 VN07 VN10; do # VN03 VN05 VN07 
+      	    python3 h5_2_tiff.py $H5FILE Rs_${B} $H5FILE.$B.tif $GCP_INTERVAL
+	    gdalwarp $WARPOPT $H5FILE.$B.tif GCOM-C/$RES/$TILE/$(basename $H5FILE).$B.tif
+	done
+    fi
+
+    #H5FILE=$WORKDIR/GC1SG1_${YYYY}${MM}${DD}*${TILE}_L2SG_NWLRQ_2001.h5
+    #FTP="ftp://ftp.gportal.jaxa.jp/standard/GCOM-C/GCOM-C.SGLI/L2.OCEAN.NWLR/2"
+
 
     # Not stable to use resampled products.
     #VGI_H5FILE=$WORKDIR/GC1SG1_${YYYY}${MM}${DD}D01D_T${TILE}_L2SG_VGI_Q_2000.h5
@@ -47,7 +51,7 @@ export -f getRSRF
 #parallel getRSRF ::: {366..730}
 #parallel getRSRF ::: {0..6} #314
 #parallel --bar getRSRF ::: {0..14}
-parallel getRSRF ::: {0..315}
+parallel getRSRF ::: {0..331}
 #for i in {268..273}; do getRSRF $i; done
 
 for B in VN04 VN06 VN07 VN10 ; do #
